refactor(search-service): delete barrel file and update imports across frontend
