+++ public/app/core/components/Select/SortPicker.tsx
@@ -3,7 +3,7 @@ import { useAsync } from 'react-use';
 import { SelectableValue } from '@grafana/data';
 import { Icon, Select } from '@grafana/ui';
 import { DEFAULT_SORT } from 'app/features/search/constants';
-import { getGrafanaSearcher } from 'app/features/search/service';
+import { getGrafanaSearcher } from 'app/features/search/service/searcher';
 
 export interface Props {
   onChange: (sortValue: SelectableValue) => void;
