+++ public/app/features/playlist/PlaylistForm.tsx
@@ -8,7 +8,7 @@ import { DashboardPicker } from 'app/core/components/Select/DashboardPicker';
 import { TagFilter } from 'app/core/components/TagFilter/TagFilter';
 import { Trans, t } from 'app/core/internationalization';
 
-import { getGrafanaSearcher } from '../search/service';
+import { getGrafanaSearcher } from '../search/service/searcher';
 
 import { PlaylistTable } from './PlaylistTable';
 import { Playlist } from './types';
