+++ public/app/features/playlist/api.ts
@@ -10,7 +10,8 @@ import { dispatch } from 'app/store/store';
 
 import { ScopedResourceClient } from '../apiserver/client';
 import { Resource, ResourceForCreate, ResourceClient } from '../apiserver/types';
-import { DashboardQueryResult, getGrafanaSearcher, SearchQuery } from '../search/service';
+import { getGrafanaSearcher } from '../search/service/searcher';
+import { DashboardQueryResult, SearchQuery } from '../search/service/types';
 
 import { Playlist, PlaylistItem, PlaylistAPI } from './types';
