--- public/app/features/search/service/index.ts
@@ -1,2 +1,0 @@
-export * from './types';
-export { getGrafanaSearcher } from './searcher';
