+++ public/app/plugins/datasource/grafana/components/SearchEditor.tsx
@@ -2,7 +2,7 @@ import { useEffect, useState } from 'react';
 import * as React from 'react';
 
 import { InlineField, Input, InlineFieldRow, CodeEditor } from '@grafana/ui';
-import { SearchQuery } from 'app/features/search/service';
+import { SearchQuery } from 'app/features/search/service/types';
 
 interface Props {
   value: SearchQuery;
